# src/main/resources/api-docs.yaml
openapi: 3.0.3
info:
  title: Financial Management API
  description: |
    A comprehensive REST API for financial management system that handles users, accounts, and transactions.
    
    ## Features
    - User management (CRUD operations)
    - Account management with different account types
    - Transaction processing and history
    - Real-time balance updates
    - Transaction status tracking
    
    ## Authentication
    Currently, this API does not require authentication. In production, implement proper security measures.
    
    ## Rate Limiting
    API calls are limited to 1000 requests per hour per IP address.
    
  version: 1.0.0
  contact:
    name: Financial API Support
    email: support@financial.com
    url: https://financial.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081/api
    description: Development server
  - url: https://api.financial.com/v1
    description: Production server
...
paths:
...
  /accounts/{id}:
    get:
      tags:
        - Accounts
      summary: Get account by ID
      description: Retrieve a specific account by its ID
      operationId: getAccountById
      parameters:
        - name: id
          in: path
          required: true
          description: Account ID
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Account found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /transactions:
    get:
      tags:
        - Transactions
      summary: Get all transactions
      description: Retrieve a list of all transactions
      operationId: getAllTransactions
      parameters:
        - name: accountId
          in: query
          description: Filter by account ID
          schema:
            type: integer
            format: int64
        - name: transactionType
          in: query
          description: Filter by transaction type
          schema:
            $ref: '#/components/schemas/TransactionType'
        - name: status
          in: query
          description: Filter by transaction status
          schema:
            $ref: '#/components/schemas/TransactionStatus'
        - name: fromDate
          in: query
          description: Filter transactions from this date
          schema:
            type: string
            format: date-time
        - name: toDate
          in: query
          description: Filter transactions to this date
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          description: Page number (0-based)
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Successfully retrieved transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  pageable:
                    $ref: '#/components/schemas/Pageable'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Transactions
      summary: Create a new transaction
      description: Process a new transaction
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Transaction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Bad request - validation errors or insufficient funds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /transactions/{id}:
    get:
      tags:
        - Transactions
      summary: Get transaction by ID
      description: Retrieve a specific transaction by its ID
      operationId: getTransactionById
      parameters:
        - name: id
          in: path
          required: true
          description: Transaction ID
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Transaction found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Account:
      type: object
      required:
        - id
        - userId
        - accountNumber
        - accountType
        - balance
        - currency
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the account
          example: 1
        userId:
          type: integer
          format: int64
          description: ID of the user who owns this account
          example: 1
        accountNumber:
          type: string
          maxLength: 20
          description: Unique account number
          example: "ACC001"
        accountType:
          $ref: '#/components/schemas/AccountType'
        balance:
          type: number
          format: decimal
          description: Current account balance
          example: 1500.00
        currency:
          type: string
          maxLength: 3
          description: Account currency code
          example: "USD"
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2023-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Account last update timestamp
          example: "2023-01-01T00:00:00Z"
...
    AccountType:
      type: string
      enum:
        - CHECKING
        - SAVINGS
        - CREDIT
        - INVESTMENT
      description: Type of account
      example: "CHECKING"
    Transaction:
      type: object
      required:
        - id
        - amount
        - transactionType
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the transaction
          example: 1
        fromAccountId:
          type: integer
          format: int64
          nullable: true
          description: ID of the source account (null for deposits)
          example: 1
        toAccountId:
          type: integer
          format: int64
          nullable: true
          description: ID of the destination account (null for withdrawals)
          example: 2
        amount:
          type: number
          format: decimal
          description: Transaction amount
          example: 200.00
        transactionType:
          $ref: '#/components/schemas/TransactionType'
        description:
          type: string
          description: Transaction description
          example: "Transfer to savings"
        status:
          $ref: '#/components/schemas/TransactionStatus'
        createdAt:
          type: string
          format: date-time
          description: Transaction creation timestamp
          example: "2023-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Transaction last update timestamp
          example: "2023-01-01T00:00:00Z"
    CreateTransactionRequest:
      type: object
      required:
        - amount
        - transactionType
      properties:
        fromAccountId:
          type: integer
          format: int64
          description: ID of the source account (required for withdrawals and transfers)
          example: 1
        toAccountId:
          type: integer
          format: int64
          description: ID of the destination account (required for deposits and transfers)
          example: 2
        amount:
          type: number
          format: decimal
          minimum: 0.01
          description: Transaction amount
          example: 200.00
        transactionType:
          $ref: '#/components/schemas/TransactionType'
        description:
          type: string
          description: Transaction description
          example: "Transfer to savings"
    TransactionType:
      type: string
      enum:
        - DEPOSIT
        - WITHDRAWAL
        - TRANSFER
        - PAYMENT
      description: Type of transaction
      example: "TRANSFER"
    TransactionStatus:
      type: string
      enum:
        - PENDING
        - COMPLETED
        - FAILED
        - CANCELLED
      description: Status of the transaction
      example: "COMPLETED"
    Pageable:
      type: object
      properties:
        pageNumber:
          type: integer
          description: Page number (0-based)
          example: 0
        pageSize:
          type: integer
          description: Number of items per page
          example: 20
        sort:
          type: object
          properties:
            sorted:
              type: boolean
              example: true
            unsorted:
              type: boolean
              example: false
            empty:
              type: boolean
              example: false
...
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2023-01-01T00:00:00Z"
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type
          example: "Bad Request"
        message:
          type: string
          description: Error message
          example: "Validation failed"
        path:
          type: string
          description: Request path
          example: "/api/users"
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            validationErrors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: "email"
                  message:
                    type: string
                    example: "Email format is invalid"
                  rejectedValue:
                    type: string
                    example: "invalid-email"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
security:
  - ApiKeyAuth: []
  - BearerAuth: []